backend/requirements.txt
fastapi
httpx
pymongo
dotenv
fastapi-cache2
